refactor(lib): move run related functions into run module
